requirements.txt
a==1.0
alabaster==0.7.12
asgi-redis==1.4.3
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
bootstrap-py==1.0.1
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.6
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
docutils==0.16
gitdb==4.0.5
GitPython==3.1.2
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
incremental==17.5.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack-python==0.5.6
nodeenv==1.3.5
nodejs==0.1.1
npm==0.1.1
optional-django==0.1.0
packaging==20.3
pguard==0.4.0
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytz==2020.1
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
six==1.14.0
smmap==3.0.4
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
